Create new `h.schemas.forms.accounts.login` module and tests
Update `h.views.accounts` to use new `LoginSchema`

Remove `LoginSchema` from `h.accounts.schemas`
