+++ h/accounts/schemas.py
@@ -9,7 +9,6 @@ from itsdangerous import BadData, SignatureExpired
 from jinja2 import Markup
 
 from h import i18n, models, validators
-from h.services.user import UserNotActivated
 from h.models.user import (
     EMAIL_MAX_LENGTH,
     USERNAME_MAX_LENGTH,
@@ -131,49 +130,6 @@ def _privacy_accepted_message():
     )
 
     return privacy_msg
-
-
-class LoginSchema(CSRFSchema):
-    username = colander.SchemaNode(
-        colander.String(),
-        title=_('Username / email'),
-        widget=deform.widget.TextInputWidget(autofocus=True),
-    )
-    password = colander.SchemaNode(
-        colander.String(),
-        title=_('Password'),
-        widget=deform.widget.PasswordWidget()
-    )
-
-    def validator(self, node, value):
-        super(LoginSchema, self).validator(node, value)
-
-        request = node.bindings['request']
-        username = value.get('username')
-        password = value.get('password')
-
-        user_service = request.find_service(name='user')
-        user_password_service = request.find_service(name='user_password')
-
-        try:
-            user = user_service.fetch_for_login(username_or_email=username)
-        except UserNotActivated:
-            err = colander.Invalid(node)
-            err['username'] = _("Please check your email and open the link "
-                                "to activate your account.")
-            raise err
-
-        if user is None:
-            err = colander.Invalid(node)
-            err['username'] = _('User does not exist.')
-            raise err
-
-        if not user_password_service.check_password(user, password):
-            err = colander.Invalid(node)
-            err['password'] = _('Wrong password.')
-            raise err
-
-        value['user'] = user
 
 
 class ForgotPasswordSchema(CSRFSchema):
