+++ h/schemas/forms/accounts/__init__.py
@@ -3,8 +3,10 @@
 from __future__ import unicode_literals
 
 from h.schemas.forms.accounts.edit_profile import EditProfileSchema
+from h.schemas.forms.accounts.login import LoginSchema
 
 
 __all__ = (
     'EditProfileSchema',
+    'LoginSchema',
 )
